+++ src/Message.jsx
@@ -1,7 +1,6 @@
 const Notification = require('rc-notification');
 const classnames = require('classnames');
 const React = require('react');
-const ReactDOM = require('react-dom');
 
 const defaultDuration = 1.5;
 let messageInstance;
@@ -10,36 +9,13 @@ let prefixCls = 'kuma-message';
 let transitionName = 'message-moveUp';
 let className;
 let getContainer;
+let multipleInstance = true;
 
-Notification.newInstance = function newNotificationInstance(properties) {
-  const { getContainer, ...props } = properties || {};
-  let div;
-  if (getContainer) {
-    div = getContainer();
-  } else {
-    div = document.createElement('div');
-    document.body.appendChild(div);
+function createMessageInstance() {
+  if (messageInstance && messageInstance.destroy) {
+    messageInstance.destroy();
   }
-  /* eslint-disable react/no-render-return-value */
-  const notification = ReactDOM.render(<Notification {...props} />, div);
-  /* eslint-enable react/no-render-return-value */
-  return {
-    notice(noticeProps) {
-      notification.add(noticeProps);
-    },
-    removeNotice(index) {
-      notification.remove(index);
-    },
-    component: notification,
-    destroy() {
-      ReactDOM.unmountComponentAtNode(div);
-      document.body.removeChild(div);
-    },
-  };
-};
-
-function getMessageInstance() {
-  messageInstance = messageInstance || Notification.newInstance({
+  messageInstance = Notification.newInstance({
     prefixCls,
     className,
     transitionName,
@@ -58,8 +34,7 @@ function notice(content, duration = defaultDuration, type, onClose) {
     error: 'kuma-icon kuma-icon-error',
     loading: 'kuma-loading',
   })[type];
-
-  const instance = getMessageInstance();
+  const instance = multipleInstance && messageInstance ? messageInstance : createMessageInstance();
   instance.notice({
     key,
     duration,
@@ -103,10 +78,16 @@ module.exports = {
   loading(content, duration, onClose) {
     return notice(content, duration, 'loading', onClose);
   },
+  clear() {
+    createMessageInstance();
+  },
   config(options) {
-    prefixCls = options.prefixCls || prefixCls;
-    transitionName = options.transitionName || transitionName;
-    className = options.className || className;
-    getContainer = options.getContainer;
+    if (options) {
+      prefixCls = options.prefixCls || prefixCls;
+      transitionName = options.transitionName || transitionName;
+      className = options.className || className;
+      multipleInstance = options.multipleInstance !== false;
+      getContainer = options.getContainer;
+    }
   },
 };